Revert "."
This reverts commit 0aa47d7c3c2e3b3693b19db4b0d926d52325a94a.

+++ Src/mcp23017.h
@@ -66,7 +66,7 @@ void updateRegisterBit(uint8_t, uint8_t, uint8_t, uint8_t);
 uint8_t readRegister(uint8_t);
 
 //Functions prototypes for hardware abstraction
-void begin(I2C_HandleTypeDef * ,uint16_t);
+void begin(I2C_HandleTypeDef * ,uint8_t);
 void pinMode(uint8_t, uint8_t);
 void pullUp(uint8_t, uint8_t);
 uint8_t digitalRead(uint8_t);
